refactor(#174): types 리팩토링에 따라 import 경로 및 문구 변경
