+++ frontend/src/app/(servers)/components/ChannelDropdownMenu.tsx
@@ -1,10 +1,8 @@
 "use client";
 
 import React from "react";
-import { Channel } from "@/app/(server-setup)/types/Channel";
-import { ChannelKind } from "@/app/(server-setup)/types/ChannelKind";
-import { AccessType } from "@/app/(server-setup)/types/AccessType";
-import { ChannelRole } from "@/app/(server-setup)/types/ChannelRole";
+import type { Channel } from "@/app/(server-setup)/types";
+import { ChannelKind, AccessType, ChannelRole } from "@/app/(server-setup)/types";
 import { useChannelFlow } from "@/app/(server-setup)/hooks/useChannelFlow";
 import { useGetUserInfoQuery } from "@/app/(auth)/hooks/useAuthMutations";
